+++ tests/test_profile_builder.py
@@ -248,7 +248,7 @@ def test_extension_codelists(caplog):
 
         # Combines codelist additions and removals.
         assert len(plus_party_role) == 13
-        assert plus_party_role[-1]['Code'] == 'enquirer'
+        assert sorted(plus_party_role)[-1]['Code'] == 'socialWitness'
 
         # Logs ignored codelists.
         assert len(caplog.records) == 1